requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.2.0
cvzone==1.6.1
cycler==0.12.1
Flask==3.1.2

flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
namex==0.1.0
numpy==1.23.5
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.10.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.5.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2

